jumping fixed
add a jump function to make it work properly I used the character controller components standard isGrounded check. and remove my own ground check because no reason to use both.

Next Ill make the gravity function to the EventManager so anything that needs it can inherit from it and not write the same function multiple times